+++ backend/src/main/java/org/prgrms/devconnect/api/service/board/BoardService.java
@@ -2,19 +2,17 @@ package org.prgrms.devconnect.api.service.board;
 
 import lombok.RequiredArgsConstructor;
 import org.prgrms.devconnect.api.controller.board.dto.request.BoardCreateRequestDto;
-import org.prgrms.devconnect.api.service.member.MemberService;
 import org.prgrms.devconnect.common.exception.ExceptionCode;
 import org.prgrms.devconnect.common.exception.board.BoardException;
 import org.prgrms.devconnect.domain.define.board.entity.Board;
 import org.prgrms.devconnect.domain.define.board.entity.BoardTechStackMapping;
 import org.prgrms.devconnect.domain.define.board.repository.BoardRepository;
 import org.prgrms.devconnect.domain.define.jobpost.entity.JobPost;
-import org.prgrms.devconnect.domain.define.jobpost.repository.JobPostRepository;
+import org.prgrms.devconnect.domain.define.jobpost.entity.repository.JobPostRepository;
 import org.prgrms.devconnect.domain.define.member.entity.Member;
 import org.prgrms.devconnect.domain.define.member.repository.MemberRepository;
 import org.prgrms.devconnect.domain.define.techstack.entity.TechStack;
 import org.prgrms.devconnect.domain.define.techstack.repository.TechStackRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -28,7 +26,9 @@ public class BoardService {
 
   private final BoardRepository boardRepository;
 
-  private final MemberService memberService;
+  //TODO 수정 필요
+  private final MemberRepository memberRepository;
+//  private final MemberService memberService;
 
   private final JobPostRepository jobPostRepository;
 
@@ -37,7 +37,9 @@ public class BoardService {
 
   @Transactional
   public Long createBoard(BoardCreateRequestDto boardCreateRequestDto) {
-    Member member = memberService.findMemberById(boardCreateRequestDto.memberId());
+    //TODO 수정 필요
+//    Member member = memberService.findMemberById(boardCreateRequestDto.memberId());
+    Member member = memberRepository.findById(boardCreateRequestDto.memberId()).orElseThrow(() -> new RuntimeException("Member Not Found"));
 
     // JobPost 존재 여부 확인. 없을 경우 null로 처리
     JobPost jobPost = null;
