+++ backend/src/main/java/org/prgrms/devconnect/domain/define/chatting/entity/ChatParticipation.java
@@ -12,8 +12,6 @@ import org.prgrms.devconnect.domain.define.member.entity.Member;
 import java.util.ArrayList;
 import java.util.List;
 
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Table(name = "chat_participation")
