+++ backend/src/main/java/org/prgrms/devconnect/domain/define/chatting/entity/repository/ChatParticipationRepository.java
@@ -1,6 +1,7 @@
-package org.prgrms.devconnect.chatting.repository;
+package org.prgrms.devconnect.domain.define.chatting.entity.repository;
 
-import org.prgrms.devconnect.chatting.entity.ChatParticipation;
+
+import org.prgrms.devconnect.domain.define.chatting.entity.ChatParticipation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
