+++ backend/src/main/java/org/prgrms/devconnect/domain/define/member/entity/Member.java
@@ -1,17 +1,5 @@
 package org.prgrms.devconnect.domain.define.member.entity;
 
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import java.util.ArrayList;
-import java.util.List;
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -19,10 +7,8 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.prgrms.devconnect.domain.define.Timestamp;
+import org.prgrms.devconnect.domain.define.chatting.entity.ChatParticipation;
 import org.prgrms.devconnect.domain.define.member.entity.constant.Interest;
-import org.prgrms.devconnect.chatting.entity.ChatParticipation;
-import org.prgrms.devconnect.common.audit.Timestamp;
-import org.prgrms.devconnect.member.entity.constant.Interest;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -72,6 +58,10 @@ public class Member extends Timestamp {
   @Column(name = "interest", length = 100)
   private Interest interest;
 
+  @OneToMany(mappedBy = "member", fetch = FetchType.LAZY)
+  @JsonManagedReference
+  List<ChatParticipation> chattinglist = new ArrayList<>();
+
   @Builder
   public Member(String email, String password, String nickname, String job, String affiliation,
       int career, String selfIntroduction, String blogLink, String githubLink, Interest interest,
@@ -100,10 +90,6 @@ public class Member extends Timestamp {
   public boolean isValidPassword(String password) {
     return this.password.equals(password);
   }
-
-  @OneToMany(mappedBy = "member", fetch = FetchType.LAZY)
-  @JsonManagedReference
-  List<ChatParticipation> chattinglist = new ArrayList<>();
 
   public void addChattings(ChatParticipation chatParticipation){
     chattinglist.add(chatParticipation);
