+++ backend/src/main/java/org/prgrms/devconnect/domain/define/techstack/entity/TechStack.java
@@ -7,6 +7,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -27,4 +28,9 @@ public class TechStack {
   @Column(name = "job_code", length = 50)
   private String code;
 
+  @Builder
+  public TechStack(String name, String code) {
+    this.name = name;
+    this.code = code;
+  }
 }
