--- backend/src/main/java/org/prgrms/devconnect/member/repository/MemberRepository.java
@@ -1,9 +1,0 @@
-package org.prgrms.devconnect.member.repository;
-
-import org.prgrms.devconnect.member.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-}
