+++ backend/src/test/java/org/prgrms/devconnect/chatting/ChattingServiceTest.java
@@ -3,15 +3,15 @@ package org.prgrms.devconnect.chatting;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.prgrms.devconnect.chatting.entity.ChatParticipation;
-import org.prgrms.devconnect.chatting.entity.ChattingRoom;
-import org.prgrms.devconnect.chatting.entity.constant.ChattingRoomStatus;
-import org.prgrms.devconnect.chatting.repository.ChatParticipationRepository;
-import org.prgrms.devconnect.chatting.repository.ChattingRoomRepository;
-import org.prgrms.devconnect.chatting.service.ChattingService;
-import org.prgrms.devconnect.member.entity.Member;
-import org.prgrms.devconnect.member.entity.constant.Interest;
-import org.prgrms.devconnect.member.repository.MemberRepository;
+import org.prgrms.devconnect.api.service.chatting.ChattingService;
+import org.prgrms.devconnect.domain.define.chatting.entity.ChatParticipation;
+import org.prgrms.devconnect.domain.define.chatting.entity.ChattingRoom;
+import org.prgrms.devconnect.domain.define.chatting.entity.constant.ChattingRoomStatus;
+import org.prgrms.devconnect.domain.define.chatting.entity.repository.ChatParticipationRepository;
+import org.prgrms.devconnect.domain.define.chatting.entity.repository.ChattingRoomRepository;
+import org.prgrms.devconnect.domain.define.member.entity.Member;
+import org.prgrms.devconnect.domain.define.member.entity.constant.Interest;
+import org.prgrms.devconnect.domain.define.member.repository.MemberRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
@@ -19,7 +19,8 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 @SpringBootTest
 @ActiveProfiles("test")
