+++ backend/src/test/java/org/prgrms/devconnect/chatting/entity/ChattingRoomTest.java
@@ -2,7 +2,8 @@ package org.prgrms.devconnect.chatting.entity;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.prgrms.devconnect.chatting.entity.constant.ChattingRoomStatus;
+import org.prgrms.devconnect.domain.define.chatting.entity.ChattingRoom;
+import org.prgrms.devconnect.domain.define.chatting.entity.constant.ChattingRoomStatus;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
